requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
geoalchemy2==0.14.2
asyncpg==0.29.0

# Geospatial
shapely==2.0.1
geopandas==1.0.1
geojson-pydantic==1.0.1

# Data processing
pandas==2.2.3
numpy==1.26.4

# HTTP client
requests==2.32.3
httpx==0.25.2
aiofiles==23.2.1

# Utilities
pydantic==2.5.2
pyyaml==6.0.1
python-json-logger==2.0.7

# API extensions
fastapi-pagination==0.12.13
slowapi==0.1.9
fastapi-utils==0.2.1